pyproject.toml
[build-system]
requires = ["setuptools>=65.5"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-hybrid"
version = "0.1.0"
description = "Hybrid RAG ingestion UI using pgVector and Neo4j"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=5.28.1",
    "numpy>=2.2.6",
    "openai>=1.79.0",
    "pdfplumber>=0.11.7",
    "psycopg[binary]>=3.2.9",
    "pypdf>=5.5.0",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "pydantic>=2.13.3",
    "streamlit>=1.44.1",
]

[tool.setuptools.packages.find]
include = ["rag_hybrid*"]
